third_party/WebKit/LayoutTests/external/wpt/payment-request/allowpaymentrequest/allowpaymentrequest-attribute-cross-origin-bc-containers.https-expected.txt

This is a testharness.js-based test.
FAIL iframeassert_equals: expected "Success" but got "Exception"
FAIL frameassert_array_equals: property 0, expected true but got false
FAIL objectassert_array_equals: property 0, expected true but got false
FAIL embedassert_array_equals: property 0, expected true but got false
Harness: the test ran to completion.
